requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
plotly>=5.0.0
fpdf>=1.7.2
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.24
openpyxl>=3.0.10
networkx>=2.8.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.20.0 